test entry point file not found
